requirements.txt
bitsandbytes==0.39.0
braceexpand==0.1.7
einops==0.8.0
fsspec==2024.2.0
ftfy==6.2.0
fvcore==0.1.5.post20221221
# horovod==0.28.1
huggingface_hub==0.20.3
iopath==0.1.10
joblib==1.3.2
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.0.3
peft==0.3.0
Pillow==10.3.0
regex==2023.12.25
safetensors==0.4.3
scikit_learn==1.3.2
tensorboardX==2.6.2.2
timm==1.0.3
tqdm==4.66.2
transformers==4.32.1
wandb==0.16.3
webdataset==0.2.86
